# apps/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Barracks API Documentation",
    "description": "",
    "version": "1.0.0"
  },
  "host": "",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKey": {
      "type": "apiKey",
      "name": "x-api-token",
      "in": "header"
    }
  },
  "paths": {
...
    "/proxy/web/assets/img/devices/": {
      "get": {
        "tags": [
          "Proxy"
        ],
        "description": "",
        "parameters": [
          {
            "name": "devicename",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/proxy/Items/Images/Backdrop/": {
      "get": {
        "tags": [
          "Proxy"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "fillWidth",
            "in": "query",
            "type": "string"
          },
          {
            "name": "quality",
            "in": "query",
            "type": "string"
          },
          {
            "name": "blur",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/proxy/Items/Images/Primary/": {
      "get": {
        "tags": [
          "Proxy"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "fillWidth",
            "in": "query",
            "type": "string"
          },
          {
            "name": "quality",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/proxy/Users/Images/Primary/": {
      "get": {
        "tags": [
          "Proxy"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "fillWidth",
            "in": "query",
            "type": "string"
          },
          {
            "name": "quality",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/proxy/getSessions": {
      "get": {
        "tags": [
          "Proxy"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/proxy/getAdminUsers": {
      "get": {
        "tags": [
          "Proxy"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/proxy/getRecentlyAdded": {
      "get": {
        "tags": [
          "Proxy"
        ],
        "description": "",
        "parameters": [
          {
            "name": "libraryid",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/proxy/validateSettings": {
      "post": {
        "tags": [
          "Proxy"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "url": {
                  "example": "any"
                },
                "apikey": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/getRecentlyAdded": {
      "get": {
        "tags": [
          "API"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-api-token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "req",
            "in": "query",
            "type": "string"
          },
          {
            "name": "libraryid",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "GroupResults",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "apiKey": []
    }
  ]
}